+++ src/main/java/teamkiim/koffeechat/domain/comment/service/CommentService.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import teamkiim.koffeechat.domain.aescipher.AESCipher;
+import teamkiim.koffeechat.global.aescipher.AESCipher;
 import teamkiim.koffeechat.domain.comment.domain.Comment;
 import teamkiim.koffeechat.domain.comment.dto.request.CommentServiceRequest;
 import teamkiim.koffeechat.domain.comment.dto.request.ModifyCommentServiceRequest;
