+++ src/main/java/teamkiim/koffeechat/domain/post/common/service/PostService.java
@@ -6,7 +6,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import teamkiim.koffeechat.domain.aescipher.AESCipher;
+import teamkiim.koffeechat.global.aescipher.AESCipher;
 import teamkiim.koffeechat.domain.bookmark.domain.Bookmark;
 import teamkiim.koffeechat.domain.bookmark.repository.BookmarkRepository;
 import teamkiim.koffeechat.domain.bookmark.service.BookmarkService;
